Broke out js files to each component html file, added middleware authentication with firebase, add render pages for the ui
