+++ backend/public/portfolio.js
@@ -1,0 +1,6 @@
+//-----------GLOBAL HTML OBJECTS
+
+
+
+
+//-----------GLOBAL EVENTS
