--- main.css
@@ -1,14 +1,0 @@
-.login{
-    margin-top: 100px;
-    margin-left:500px;
-}
-
-.color{
-    color: red;
-    background-color: red;
-    
-}
-
-.hidden{
-    display: none;
-}
